+++ test/SignUp_test.go
@@ -4,15 +4,12 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/gin-gonic/gin"
 	"github.com/ribeirosaimon/motion-go/domain"
 	"github.com/ribeirosaimon/motion-go/test/util"
 )
 
-var signUpEnginer = gin.New()
-
 func TestLoginAndSignUpController(t *testing.T) {
-	session, err := util.SignUp(signUpEnginer, domain.USER, domain.ADMIN, domain.USER)
+	session, err := util.SignUp(testEnginer, domain.USER, domain.ADMIN, domain.USER)
 
 	if err != nil {
 		util.ErrorTest(err.Error())
