+++ test/init_test.go
@@ -1,0 +1,28 @@
+package test
+
+import (
+	"github.com/gin-gonic/gin"
+	"github.com/ribeirosaimon/motion-go/domain"
+	"github.com/ribeirosaimon/motion-go/test/util"
+)
+
+var (
+	MyToken     string
+	testEnginer *gin.Engine
+)
+
+func getToken() string {
+
+	token, err := util.SignUp(testEnginer, domain.USER, domain.ADMIN, domain.USER)
+	if err != nil {
+		util.ErrorTest(err.Error())
+	}
+
+	return token
+}
+
+func init() {
+	util.RemoveDatabase()
+	MyToken = getToken()
+	testEnginer = gin.New()
+}
